requirements.txt
Django==5.0.2
Pillow==10.2.0
django-crispy-forms==2.1
crispy-tailwind==0.5.0
django-taggit==5.0.1
gunicorn==21.2.0
whitenoise==6.6.0
python-dotenv==1.0.1